♻️ Refactor: 날짜 표시 함수를 utils로 분리
