+++ src/components/board/BoardCard.tsx
@@ -1,9 +1,8 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { format, differenceInMinutes, differenceInHours } from 'date-fns';
-import { ko } from 'date-fns/locale';
 import Badge from '../thunder/Badge';
 import { LuDot } from 'react-icons/lu';
+import { formatCreatedAt } from '../../utils/formatCreatedAt';
 
 interface BoardCardProps {
   id: number;
@@ -28,25 +27,6 @@ const BoardCard: React.FC<BoardCardProps> = ({
 }) => {
   const [imageLoaded, setImageLoaded] = useState(false);
   const defaultImageUrl = '/images/CuteEgg.svg';
-
-  const formatCreatedAt = (createdAt: string): string => {
-    const date = new Date(createdAt);
-    const now = new Date();
-
-    const minutesDifference = differenceInMinutes(now, date);
-    const hoursDifference = differenceInHours(now, date);
-
-    if (minutesDifference < 1) {
-      return '방금 전';
-    } else if (minutesDifference < 60) {
-      return `${minutesDifference}분 전`;
-    } else if (hoursDifference < 24) {
-      return `${hoursDifference}시간 전`;
-    } else {
-      return format(date, 'M월 d일 ', { locale: ko });
-    }
-  };
-
   const formattedCreatedAt = formatCreatedAt(createdAt);
 
   return (
@@ -66,7 +46,7 @@ const BoardCard: React.FC<BoardCardProps> = ({
           />
         </div>
         <div className="ml-4 flex h-full grow flex-col justify-between gap-1">
-          <h2 className="text-[20px] font-medium xs:text-[16px]">{title}</h2>
+          <h2 className="line-clamp-2 text-[20px] font-medium xs:text-[16px]">{title}</h2>
           <div className="flex gap-2">
             <Badge label={category} />
           </div>
